Rename unnamed_tuple to tuple
