+++ lang/parse/type.rs
@@ -10,7 +10,7 @@ impl Parse for TypeAnnotation {
 
         let s = match inner.as_rule() {
             Rule::list_type => Self(Refer::new(Type::List(ListType::parse(inner)?), pair.into())),
-            Rule::unnamed_tuple_type => Self(Refer::new(
+            Rule::tuple_type => Self(Refer::new(
                 Type::Tuple(TupleType::parse(inner)?),
                 pair.into(),
             )),
