+++ lang/value/mod.rs
@@ -10,13 +10,13 @@ mod list;
 mod named_tuple;
 
 #[macro_use]
-mod unnamed_tuple;
+mod tuple;
 mod value_error;
 mod value_list;
 
 pub use list::*;
 pub use named_tuple::*;
-pub use unnamed_tuple::*;
+pub use tuple::*;
 pub use value_error::*;
 
 pub use value_list::*;
@@ -59,7 +59,7 @@ pub enum Value {
     /// A tuple of named items.
     NamedTuple(NamedTuple),
     /// A tuple of unnamed items.
-    UnnamedTuple(UnnamedTuple),
+    Tuple(Tuple),
     /// A node in the render tree.
     Nodes(ModelNodes),
 }
@@ -221,7 +221,7 @@ impl crate::ty::Ty for Value {
             Value::Color(_) => Type::Color,
             Value::List(list) => list.ty(),
             Value::NamedTuple(named_tuple) => named_tuple.ty(),
-            Value::UnnamedTuple(unnamed_tuple) => unnamed_tuple.ty(),
+            Value::Tuple(unnamed_tuple) => unnamed_tuple.ty(),
             Value::Nodes(_) => Type::Nodes,
         }
     }
@@ -276,9 +276,7 @@ impl std::ops::Add for Value {
                 )))
             }
             // Add values of two tuples of the same length
-            (Value::UnnamedTuple(lhs), Value::UnnamedTuple(rhs)) => {
-                Ok(Value::UnnamedTuple((lhs + rhs)?))
-            }
+            (Value::Tuple(lhs), Value::Tuple(rhs)) => Ok(Value::Tuple((lhs + rhs)?)),
             (lhs, rhs) => Err(ValueError::InvalidOperator(format!("{lhs} + {rhs}"))),
         }
     }
@@ -420,7 +418,7 @@ impl std::fmt::Display for Value {
             Value::Color(c) => write!(f, "{c}"),
             Value::List(l) => write!(f, "{l}"),
             Value::NamedTuple(t) => write!(f, "{t}"),
-            Value::UnnamedTuple(t) => write!(f, "{t}"),
+            Value::Tuple(t) => write!(f, "{t}"),
             Value::Nodes(n) => n.dump(f),
         }
     }
@@ -443,7 +441,7 @@ impl std::fmt::Debug for Value {
             Self::Color(arg0) => write!(f, "Color: {arg0}"),
             Self::List(arg0) => write!(f, "List: {arg0}"),
             Self::NamedTuple(arg0) => write!(f, "NamedTuple: {arg0}"),
-            Self::UnnamedTuple(arg0) => write!(f, "UnnamedTuple: {arg0}"),
+            Self::Tuple(arg0) => write!(f, "UnnamedTuple: {arg0}"),
             Self::Nodes(arg0) => write!(f, "Nodes:\n {arg0}"),
         }
     }
